requirements.txt
blinker==1.7.0
branca==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
DateTime==5.3
Flask==3.0.0
folium==0.15.0
fonttools==4.44.3
fuzzywuzzy==0.18.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
polyline==2.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zope.interface==6.1
